+++ UAV-GCS/src/uav/gcs/communication/CommunicationMOSA.java
@@ -1,12 +1,22 @@
 package uav.gcs.communication;
 
 import com.google.gson.Gson;
+import java.awt.Image;
+import java.awt.image.BufferedImage;
 import java.io.BufferedReader;
+import java.io.ByteArrayInputStream;
+import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.net.Socket;
+import java.nio.ByteBuffer;
 import java.util.concurrent.Executors;
+import javax.imageio.ImageIO;
+import javax.swing.ImageIcon;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
 import lib.uav.module.comm.Client;
 import lib.uav.module.comm.Communication;
 import lib.uav.module.comm.Server;
@@ -32,17 +42,20 @@ import uav.gcs.struct.Drone;
  */
 public class CommunicationMOSA extends Communication implements Client{
 
+    private InputStream in;
     private final ReaderFileConfig config;
     private final Drone drone;
     private boolean isRunningPlanner;
+    private JPanel panelCameraData;
 
     /**
      * Class constructor
      * @param drone instance of the aircraft
      * @since version 4.0.0
      */
-    public CommunicationMOSA(Drone drone) {
+    public CommunicationMOSA(Drone drone, JPanel panelCameraData) {
         this.drone = drone;
+        this.panelCameraData = panelCameraData;
         this.stateCommunication = StateCommunication.WAITING;
         this.config = ReaderFileConfig.getInstance();
         this.isRunningPlanner = false;
@@ -64,6 +77,7 @@ public class CommunicationMOSA extends Communication implements Client{
                         socket = new Socket(config.getHostMOSA(), config.getPortNetworkMOSAandGCS());
                         output = new PrintWriter(socket.getOutputStream(), true);
                         input = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+                        in = socket.getInputStream();
                         System.out.println("UAV-GCS connected in MOSA");
                         break;
                     } catch (IOException ex) {
@@ -75,6 +89,8 @@ public class CommunicationMOSA extends Communication implements Client{
             }
         });
     }
+    
+    int index = 0;
 
     /**
      * Treats the data to be received
@@ -102,6 +118,40 @@ public class CommunicationMOSA extends Communication implements Client{
                                 }
                             }
                         } 
+                        if (in != null){
+                            try{
+                                byte[] sizeAr = new byte[4];
+                                in.read(sizeAr);
+                                int size = ByteBuffer.wrap(sizeAr).asIntBuffer().get();
+                                byte[] imageAr = new byte[size];
+                                in.read(imageAr);
+                                BufferedImage buffImage = ImageIO.read(new ByteArrayInputStream(imageAr));
+                                ImageIcon image = new ImageIcon(buffImage);
+                                JLabel labelImg = new JLabel(image);                              
+                                int dim = 500;
+                                int w = buffImage.getWidth();
+                                int h = buffImage.getHeight();
+                                System.out.println("w: " + w);
+                                System.out.println("h: " + h);
+                                int width = Math.min(w, dim);
+                                int height = Math.min(h, (dim*h)/w);
+                                System.out.println("width: " + width);
+                                System.out.println("height: " + height);
+                                labelImg.setIcon(new ImageIcon(image.getImage()
+                                        .getScaledInstance(width, 
+                                                height, Image.SCALE_DEFAULT)));                                
+                                labelImg.setVisible(true);
+                                panelCameraData.removeAll();
+                                panelCameraData.add(labelImg);                            
+                                ImageIO.write(buffImage, "jpg", new File("picture" + index +".jpg"));
+                                System.out.println("image received!!!");
+                                index++;
+                            }catch (NegativeArraySizeException ex){
+                                System.out.println("Warning [NegativeArraySizeException] try again");
+                            } catch (Exception ex){
+                                System.out.println("Warning [Exception] try again");
+                            } 
+                        }
                         Thread.sleep(Constants.TIME_TO_SLEEP_BETWEEN_MSG);
                     }
                 } catch (InterruptedException ex) {
@@ -112,7 +162,7 @@ public class CommunicationMOSA extends Communication implements Client{
                     System.out.println("Warning [IOException] receiveData()");
                     ex.printStackTrace();
                     stateCommunication = StateCommunication.DISABLED;
-                }
+                } 
             }
         });
     }