+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features:
 
+* Added special markers on Plots (Google Maps and Simple Map) for waypoints, buzzer, photos, video and spraying.
+* Added color legend of the routes on the UAV-GCS.
 * Added new scripts Beagle Bone and tests the softwares installed.
 * Added generic planner PathPlanner4m in UAV-MOSA.
 * Renamed the project UAV-Generic to Lib-UAV.